[gaim-migrate @ 14727] gaim_accounts_reorder() should not be using size_t for new_index. I fixed it to use gint, the type of all the g_list functions involved.
--HG--
extra : convert_revision : db0b4b8154e86e2e72d6061abbc0617a5bd18ead

+++ src/account.h
@@ -815,7 +815,7 @@ void gaim_accounts_delete(GaimAccount *account);
  * @param account   The account to reorder.
  * @param new_index The new index for the account.
  */
-void gaim_accounts_reorder(GaimAccount *account, size_t new_index);
+void gaim_accounts_reorder(GaimAccount *account, gint new_index);
 
 /**
  * Returns a list of all accounts.
